Wadi Gabgaba Area added
The area in and to the immediate east of the Wadi Gabgaba was famed since Pharaonic times as a source of gold. While during the Kerman Kush and Napatan Kush period this area appears to have been under more direct control of the Kushites, during the Meroitic period, it appears this region regained its autonomy; the Blemmyes tribes traded gold, metals and animal furs traded to Meroitic Kush and Ptolemaic/Roman Egypt.