Wadi Muqaddam added
The Wadi Muqaddam is thought to be an old channel of the Nile river during the Neolithic. During the post-Kerman period, Napatan period, and most especially the Meroitic period, the Wadi Muqaddam saw a number of settlements, well sites, burials and scattered pottery finds along the entire length of the wadi attest to the importance of this route. Archaeology of the past 20 years has highlighted important settlements such as Meragh and the Hosh area as clearly Meroitic.