Flags for Oreoes and Elatea
Can someone look at 02_aresia_missions_01.txt on the arseia branch when they get a chance? Only the first task shows for some reason; this was not always the case, but I am unaware as to what has caused it.  Additionally, the tooltip for that first task does not appear, although it definitely did at some point in the past. I have messed around with it a ton, but can't seem to figure out how to fix this; I don't see any obvious issues, but I'm still relatively inexperienced.