WIP Phoenicia Content
New (unfinished) mission tree. As of now, it incorporates very little of the mission tree made for the GOTN mod
The vanilla setup Phoenicia (regarding the extent of the Phoenician cities holdings) is really inaccurate
I relied heavily on the work of historian Josette Elayi.
There are a few ambiguous details which could pertain to the setup.
- Sarpeta is currently owned by Tyre, but this could have been the "territory adjacent to the city [of Sidon]" that Alexander gave to the Sidonians.
- Myriandos might've been ruled/heavily influenced by Arados during the period, but we have no direct confirmation of this
- Tripolis should be a punic republic and have possession of the territory to the north of Tripoli; however, I have not implemented it yet
The changes I made on the characters branch should be merged into dev before I start adding plausible rulers (we know the sons of some Phoenician kings during Alexander's time and its likely that they succeeded their fathers.