Fixes
It now shouldn't be possible for multiple Upper Egyptian revolts to exist at the same time (as has happened to me)
The upper egypt event now shows the ruler of the revolt and goes to the capital. Both variants now use the switch country effect. The generic one now uses the flag. Upper Egypt now uses the completed Egyptian heritage instead of the unlocalized  waset_heritage. I intend on writing a proper heritage for it at a later point.
An old  unfinished cyprus decision was accidently added when I copied over my formables. This has been deleted. A much better (and more accurate) version of the concept is present on the Diaodchi branch.