Religion Rework, Cleaning & New Culture Group (Iberia)
Cleaning and organizing the files related to the Iberic, Marite and Druidic religions (Adaptations in Hellenic and Canaanite too to reflect the changes to Iberic), as well as adding the Indo-Iberian Culture Group (Provisional Name). Rework of the names and dynasties of Iberian Culture Group. Addition of pops and cities in the Lusitanian and Vaccaei regions and the character of Aletes. And addition to some special syncretic deities (still WiP).