Update to Aletes Adventure and Expansion to the Astarte Priesthood mechanic from Turdetanian Missions
I won't be able to mod in the next weeks because my laptop will be sent to fix, so I commit another WiP of Aletes and an Update to the Priesthood thing from the Turdetanian Missions. Aletes Adventure is nearly done, just need more flavour, but it is already playable. The expansion to the Priesthood is playable too, but there is still some nasty bugs I don't like.