Ligurian name list
Many of the names are dubious. Inscriptions which *may* be Ligurian are extremely scant, however I've tried to gather what seems to be more reliable.  All in all, it's very imperfect, but is more accurate that Latin names.

For families, as only 2 are (maybe) mentioned in inscriptions, I've mostly used names of Ligurian tribes I found online. And for half of the female names I've placed Lepontic names (excluding Cisalpine Gallic) names for now. Both measures are temporary until more inscriptions can be uncovered and interpreted.

More specifics are commented in the 00_latin culture file.